Iteration 2.0!!!
Don't have a clear list this time because I introduced and got caught up in some bugs which took me long enough to deal with that I forgot the entirely of what I did.

But in general, it now MOSTLY works as intended. Still a few midpoints not functioning properly in the z axis, but for the most part, the code does what it was intended to do during runtime. Will fix the remaining bugs, then move on to the next step: Adding back in flat-topped functionality... and then maybe adding a few quality of life/ease of use things (scriptable tile that shows its hight) before moving on to making it a non-runtime tool.