Iteration 2.1
General Cleanup of Code
Fixed last known bug
Added 2nd (flat-topped) Grid to use for flat-topped hexagon layouts, the next feature I intend to implement.
Renamed a few things in Unity,
Added a reference to a GameObject to make the debugging spheres children of said object... which should be set as the object that everything belongs to. (I'm wording this poorly...)