Iteration 2.5
Struggled but still failed to get flat-topped hexagons working; I just can't get it to look at the right things for a tile's neighbors....

Added a CustomEditor script:
It has 2 buttons, "Generate Mesh," and "Clear" which build the mesh in editor and destroy it respectively.

I think I'll give up on the flat-topped hexagons for now and next work on applying sprites/materials/textures to the mesh from the tiles.